コンポーネント名の改善および掲示板リストをリクエストする関数のネスト改善
