+++ src/components/Main.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Routes, Route } from 'react-router-dom';
-import ThreadList from './ThreadList';
+import ThreadListPage from './ThreadListPage';
 import CreateThread from './CreateThread';
 import "../styles/Main.css"
 
@@ -9,7 +9,7 @@ const Main = () => {
     <main>
       <Routes>
         {/* スレッド一覧ページ */}
-        <Route path="/" element={<ThreadList />} />
+        <Route path="/" element={<ThreadListPage />} />
         {/* スレッド作成ページ */}
         <Route path="threads/new" element={<CreateThread />} />
       </Routes>
